Fix ocaml-plplot build for GCC 14 and later.
New versions of GCC promote several warnings to errors by default. In
this repository, code included from the underlying plplot C library
triggers the error
[incompatible-pointer-types](https://gcc.gnu.org/gcc-14/porting_to.html#incompatible-pointer-types).

This issue is currently breaking ocaml-plplot tests in opam's CI; for
more info please see
https://github.com/ocaml/opam-repository/issues/28095.

The GCC behavior change can be reversed by adding an extra flag, but
we must take care to verify that the C compiler in the build
environment is new enough to recognize the flag. This PR adds the
detection logic via Dune Configurator.

In addition, the PR fixes one benign logic bug at the bottom of the
same module. When the Configurator fails to find the plplot library
dependency, it should not write either `c_flags.sexp` or
`c_library_flags.sexp` to disk. But in the existing code, the call
that writes `c_library_flags.sexp` is beyond the end of the
if/then/else block and happens even in the failure case (or rather,
it *would* happen if the call to `failwith` did not make this code
path inaccessible).

Testing:

In a hermetic sandbox environment, ocaml-plplot was compiled via opam
using a recent GCC (v14) and a less recent one (v10).

Additional tests using Clang (v16 and v20) showed that no new flags
are needed in these cases.